feat: account features OK

+++ components/data-table.tsx
@@ -27,6 +27,7 @@ import {
 import { Button } from "./ui/button";
 import { Input } from "@/components/ui/input"
 import { Trash } from "lucide-react";
+import useConfirm from "@/hooks/use-confirm";
 
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[];
@@ -43,6 +44,11 @@ export function DataTable<TData, TValue>({
   onDelete,
   disabled
 }: DataTableProps<TData, TValue>) {
+
+  const [ConfirmDialog, confirm] = useConfirm(
+    "Are you sure?",
+    "You are about to perform a bulk delete."
+  )
 
   const [sorting, setSorting] = React.useState<SortingState>([])
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>([]);
@@ -67,6 +73,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmDialog/>
       <div className="flex items-center py-4">
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -82,10 +89,13 @@ export function DataTable<TData, TValue>({
             disabled={disabled}
             variant="outline"
             className="ml-auto font-normal text-xs"
-            onClick={() =>
+            onClick={async () =>
              {
-                onDelete(table.getFilteredSelectedRowModel().rows);
-                table.resetRowSelection();
+                const ok = await confirm();
+                if (ok) {
+                  onDelete(table.getFilteredSelectedRowModel().rows);
+                  table.resetRowSelection();
+                }
              }
             }
           >
